+++ source/infra/test/snapshot.test.ts
@@ -82,17 +82,6 @@ test('spoke stack synth doesnt crash', () => {
   expect(template).toMatchSnapshot();
 });
 
-test('solution doesnt crash', () => {
-  // GIVEN
-  process.env.SOLUTION_VERSION = '1.0.0';
-  process.env.SOLUTION_NAME = 'Account Assessment for AWS Organisations';
-  process.env.DIST_OUTPUT_BUCKET = 'solutions-features';
-  process.env.SOLUTION_TRADEMARKEDNAME = 'account-assessment-for-aws-organizations';
-
-  // THEN
-  expect(require("../bin/account-assessment-solution").APP_REGISTER).toBeTruthy();
-});
-
 
 function overwriteS3Keys(obj: any, value: string = 'foo.zip'): void {
   if (Array.isArray(obj)) {